requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
openai==1.3.7
spotipy==2.22.1
music21==9.1.0
mido==1.3.2
requests==2.31.0
